chore(analysis): split files
